refactor: extract shared session status runtime
