+++ src/status/status-message.runtime.ts
@@ -1,0 +1,3 @@
+export async function loadStatusMessageRuntimeModule() {
+  return await import("../auto-reply/status.runtime.js");
+}
