+++ src/status/status-queue.runtime.ts
@@ -1,0 +1,1 @@
+export { getFollowupQueueDepth, resolveQueueSettings } from "../auto-reply/reply/queue.js";
