+++ src/status/status-subagents.runtime.ts
@@ -1,0 +1,3 @@
+export { listControlledSubagentRuns } from "../agents/subagent-control.js";
+export { countPendingDescendantRuns } from "../agents/subagent-registry.js";
+export { buildSubagentsStatusLine } from "../auto-reply/reply/commands-status-subagents.js";
